+++ webview/src/lib/components/PsButton.svelte
@@ -4,6 +4,12 @@
   export let type: 'button' | 'submit' | 'reset' = 'button';
 </script>
 
-<button class={`ps-btn ${variant === 'primary' ? 'ps-btn--primary' : ''} ${variant === 'accent' ? 'ps-btn--accent' : ''} ${variant === 'ghost' ? 'ps-btn--ghost' : ''} ${variant === 'link' ? 'ps-btn--link' : ''}`} {disabled} {type}>
+<button
+  on:click
+  class={`ps-btn ${variant === 'primary' ? 'ps-btn--primary' : ''} ${variant === 'accent' ? 'ps-btn--accent' : ''} ${variant === 'ghost' ? 'ps-btn--ghost' : ''} ${variant === 'link' ? 'ps-btn--link' : ''}`}
+  type={type}
+  disabled={disabled}
+  {...$$restProps}
+>
   <slot></slot>
 </button>